git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/AudioEngine@34707 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ tools/EventClients/Clients/WiiRemote/CWIID_WiiRemote.cpp
@@ -35,9 +35,9 @@ CPacketHELO *g_Ping = NULL;
 #endif
 std::string g_BluetoothIconPath = std::string(ICON_PATH) + std::string("/bluetooth.png");
 
-long getTicks()
+int32_t getTicks()
 {
-  long ticks;
+  int32_t ticks;
   struct timeval now;
   gettimeofday(&now, NULL);
   ticks = now.tv_sec * 1000l;
